+++ dbot/asgi.py
@@ -6,15 +6,13 @@ It exposes the ASGI callable as a module-level variable named ``application``.
 For more information on this file, see
 https://docs.djangoproject.com/en/3.1/howto/deployment/asgi/
 """
-
 import os
 
-from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
-from channels.security.websocket import AllowedHostsOriginValidator
 from django.core.asgi import get_asgi_application
 
 import stt.routing
+
 
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'dbot.settings')
